requirements.txt
fastapi==0.115.11
httpx==0.28.1
numpy==2.2.3
openai==1.65.2
python-dotenv==1.0.1
scikit_learn==1.6.1
sentence_transformers==3.4.1

